Improved build.gradle file, and fixed a few Javadoc errors

+++ src/main/java/net/dark_roleplay/library/experimental/guis/IGuiElement.java
@@ -14,24 +14,24 @@ import net.minecraft.client.renderer.Tessellator;
 import net.minecraft.client.renderer.vertex.DefaultVertexFormats;
 
 public interface IGuiElement {
-	
+
 	/**
 	 * @param child
 	 * @return child ID (-1 for child was not added)
 	 */
 	public int addChild(IGuiElement child);
-	
+
 	public IGuiElement getChild(int id);
 	public void setChild(int id, IGuiElement newChild);
-	
+
 	public List<IGuiElement> getChildren();
-	
+
 	public void draw(int mouseX, int mouseY, float partialTicks);
-	
+
 	public boolean mouseClicked(int mouseX, int mouseY, int mouseButton);
 	public boolean mousePressed(Minecraft mc, int mouseX, int mouseY);
 	public void mouseReleased(int mouseX, int mouseY);
-	
+
 	public int getPosX();
 	public int getPosY();
 	public IGuiElement setSize(int width, int height);
@@ -39,28 +39,28 @@ public interface IGuiElement {
 	public int getWidth();
 	public int getHeight();
 	public void setPos(int posX, int posY);
-	
+
 	public boolean isVisible();
 	public void setVisible(boolean visible);
-	
+
 	public boolean isEnabled();
 	public void setEnabled(boolean enabled);
-	
+
 	public void mouseDragged(Minecraft mc, int mouseX, int mouseY);
 	public boolean isHovered(int mouseX, int mouseY);
-	
+
 	public void handleMouseInput() throws IOException;
-	
+
 	public abstract class IMPL extends Gui implements IGuiElement{
 
 		protected int posX = 0, posY = 0;
 		protected int width = 10, height = 10;
 		protected boolean visible = true;
 		protected boolean enabled = true;
-		
-		
+
+
 		protected List<IGuiElement> children = new ArrayList<IGuiElement>();
-		
+
 		@Override
 		public int addChild(IGuiElement child) {
 			return -1;
@@ -68,7 +68,7 @@ public interface IGuiElement {
 
 		@Override
 		public void setChild(int id, IGuiElement newChild){
-			
+
 		}
 
 		@Override
@@ -80,7 +80,7 @@ public interface IGuiElement {
 		public int getPosY() {
 			return this.posY;
 		}
-		
+
 		@Override
 		public IGuiElement setSize(int width, int height) {
 			this.width = width;
@@ -112,19 +112,19 @@ public interface IGuiElement {
 				return null;
 			}
 		}
-		
+
 		@Override
 		public List<IGuiElement> getChildren() {
 			return this.children;
 		}
-		
+
 		@Override
 		public void mouseDragged(Minecraft mc, int mouseX, int mouseY){
 			for(IGuiElement element : this.children){
 				element.mouseDragged(mc, mouseX, mouseY);
 			}
 		}
-		
+
 		@Override
 		public boolean mouseClicked(int mouseX, int mouseY, int mouseButton){
 			if(this.children != null){
@@ -162,46 +162,46 @@ public interface IGuiElement {
 				element.mouseReleased(mouseX, mouseY);
 			}
 		}
-		
+
 		@Override
 		public void handleMouseInput() throws IOException{
 			for(IGuiElement element : this.children){
 				element.handleMouseInput();
 			}
 		}
-		
-		
+
+
 		@Override
 		public boolean isVisible(){
 			return this.visible;
 		}
-		
+
 		@Override
 		public void setVisible(boolean visible){
 			this.visible = visible;
 		}
-		
+
 		@Override
 		public boolean isEnabled(){
 			return this.enabled;
 		}
-		
+
 		@Override
 		public void setEnabled(boolean enabled){
 			this.enabled = enabled;
 		}
-		
+
 		@Override
 		public boolean isHovered(int mouseX, int mouseY){
 			return mouseX >= 0 && mouseY >= 0 && mouseX <= this.width && mouseY <= this.height;
 		}
-		
+
 		protected static void drawLine(int x1, int y1, int x2, int y2, int color) {
 
-			float f3 = (float) (color >> 24 & 255) / 255.0F;
-			float f = (float) (color >> 16 & 255) / 255.0F;
-			float f1 = (float) (color >> 8 & 255) / 255.0F;
-			float f2 = (float) (color & 255) / 255.0F;
+			float f3 = (color >> 24 & 255) / 255.0F;
+			float f = (color >> 16 & 255) / 255.0F;
+			float f1 = (color >> 8 & 255) / 255.0F;
+			float f2 = (color & 255) / 255.0F;
 			Tessellator tessellator = Tessellator.getInstance();
 			BufferBuilder vertexbuffer = tessellator.getBuffer();
 			GlStateManager.enableBlend();
@@ -210,15 +210,15 @@ public interface IGuiElement {
 			GlStateManager.color(f, f1, f2, f3);
 
 			float angle = getAngleRad(new Point(x1,y1), new Point(x2,y2));
-			
+
 			double halfPi = Math.PI / 2;
-			
+
 			float a1 = (float) (x1 + 0.5F * Math.cos(angle + halfPi)), a2 = (float) (y1 + 0.5F * Math.sin(angle + halfPi));
 			float b1 = (float) (x2 + 0.5F * Math.cos(angle + halfPi)), b2 = (float) (y2 + 0.5F * Math.sin(angle + halfPi));
 			float c1 = (float) (x2 + 0.5F * Math.cos(angle - halfPi)), c2 = (float) (y2 + 0.5F * Math.sin(angle - halfPi));
 			float d1 = (float) (x1 + 0.5F * Math.cos(angle - halfPi)), d2 = (float) (y1 + 0.5F * Math.sin(angle - halfPi));
 
-			
+
 			vertexbuffer.begin(7, DefaultVertexFormats.POSITION);
 			//new
 			vertexbuffer.pos(a1, a2, 0.0D).endVertex();
@@ -229,19 +229,19 @@ public interface IGuiElement {
 			GlStateManager.enableTexture2D();
 			GlStateManager.disableBlend();
 		}
-		
+
 		private static float getAngleRad(Point a, Point b){
 			Point vec = new Point(b.getX() - a.getX(), b.getY() - a.getY());
 			return (float) (Math.acos(vec.getX() / (Math.sqrt(Math.pow(vec.getX(), 2) + Math.pow(vec.getY(), 2)))));
 		}
-		
+
 		public static void drawCenteredRect(int centerX, int centerY, int radius, double angle, int color){
 			double rads = Math.toRadians(angle);
 			double diagRadius = radius * Math.sqrt(2);
-	        float f3 = (float)(color >> 24 & 255) / 255.0F;
-	        float f = (float)(color >> 16 & 255) / 255.0F;
-	        float f1 = (float)(color >> 8 & 255) / 255.0F;
-	        float f2 = (float)(color & 255) / 255.0F;
+	        float f3 = (color >> 24 & 255) / 255.0F;
+	        float f = (color >> 16 & 255) / 255.0F;
+	        float f1 = (color >> 8 & 255) / 255.0F;
+	        float f2 = (color & 255) / 255.0F;
 	        Tessellator tessellator = Tessellator.getInstance();
 	        BufferBuilder bufferbuilder = tessellator.getBuffer();
 	        GlStateManager.enableBlend();
@@ -257,11 +257,11 @@ public interface IGuiElement {
 	        GlStateManager.enableTexture2D();
 	        GlStateManager.disableBlend();
 	    }
-		
+
 		/**
 		 * @param posX "X" position of the top left corner
 		 * @param posY "Y" position of the top left corner
-		 * @param widht Size of the rect on the "X" axis
+		 * @param width Size of the rect on the "X" axis
 		 * @param height Size of the rect on the "Y" axis
 		 * @param u "X" position for the texture, top left corner
 		 * @param v "Y" position for the texture, top left corner
@@ -278,7 +278,7 @@ public interface IGuiElement {
 	        bufferbuilder.pos(posX, posY, this.zLevel).tex(u, v).endVertex();
 	        tessellator.draw();
 	    }
-		
+
 		public void drawTexturedCenteredRect(int centerX, int centerY, float u, float v, float uMax, float vMax, int radius, double angle){
 			double rads = Math.toRadians(angle);
 			double diagRadius = radius * Math.sqrt(2);
@@ -291,13 +291,13 @@ public interface IGuiElement {
 	        bufferbuilder.pos(centerX + radius * Math.cos(rads + Math.PI * 1.75), centerY + radius * Math.sin(rads + Math.PI * 1.75), 0.0D).tex(u, v).endVertex();
 	        tessellator.draw();
 	    }
-		
+
 		public static void drawCenteredCircle(int centerX, int centerY, int radius, int resoulution, int color){
 			double diagRadius = radius * Math.sqrt(2);
-	        float f3 = (float)(color >> 24 & 255) / 255.0F;
-	        float f = (float)(color >> 16 & 255) / 255.0F;
-	        float f1 = (float)(color >> 8 & 255) / 255.0F;
-	        float f2 = (float)(color & 255) / 255.0F;
+	        float f3 = (color >> 24 & 255) / 255.0F;
+	        float f = (color >> 16 & 255) / 255.0F;
+	        float f1 = (color >> 8 & 255) / 255.0F;
+	        float f2 = (color & 255) / 255.0F;
 	        Tessellator tessellator = Tessellator.getInstance();
 	        BufferBuilder bufferbuilder = tessellator.getBuffer();
 	        GlStateManager.enableBlend();
@@ -315,19 +315,19 @@ public interface IGuiElement {
         	for(;i < 4; i++){
 	        	bufferbuilder.pos(centerX + radius * Math.cos(Math.PI * (steps * (resoulution-i))), centerY + radius * Math.sin(Math.PI * (steps * (resoulution-i))), 0.0D).endVertex();
         	}
-        	
+
         	for(; i < resoulution; i += 2){
             	bufferbuilder.pos(startX, startY, 0.0D).endVertex();
         		bufferbuilder.pos(centerX + radius * Math.cos(Math.PI * (steps * (resoulution - i + 1))), centerY + radius * Math.sin(Math.PI * (steps * (resoulution - i + 1))), 0.0D).endVertex();
         		bufferbuilder.pos(centerX + radius * Math.cos(Math.PI * (steps * (resoulution - i))), centerY + radius * Math.sin(Math.PI * (steps * (resoulution - i))), 0.0D).endVertex();
         		bufferbuilder.pos(centerX + radius * Math.cos(Math.PI * (steps * (resoulution - i - 1))), centerY + radius * Math.sin(Math.PI * (steps * (resoulution - i - 1))), 0.0D).endVertex();
         	}
-    		
+
 	        tessellator.draw();
 	        GlStateManager.enableTexture2D();
 	        GlStateManager.disableBlend();
 	    }
-		
+
 		protected void drawTiled(int posX, int posY, int width, int height, int u, int v, int tileWidth, int tileHeight, int textureWidth, int textureHeight){
 			if(height > tileHeight && width <= tileWidth){
 				for(int i = 0; i < height; i += tileHeight){
